Change: components 내 Shared 폴더를 Web, Game 안으로 각각 이동시킴 Modal, Header는 게임용으로 따로 분리시킴
