+++ src/components/Game/Shared/Profile.jsx
@@ -1,23 +1,21 @@
 import React, { useState } from "react";
+import styled from "styled-components";
 import PropTypes from "prop-types";
 
+import { FlexBox } from "@/styles/FlexStyle";
 import Player from "./Player";
-import {
-  ProfileWrapper,
-  AvatarImage,
-  ProfileInfoWrapper,
-  ProfileInfoField,
-  ProfileInfoKey,
-  ProfileInfo,
-  ExpWrapper,
-  ExpBar,
-  ProgressToNextLevel,
-  ExpText,
-  ActiveToggle,
-  ActiveCircle
-} from "./ProfileStyle";
 
-const Profile = ({ isAdmin, profileInfos }) => {
+const init = {
+  avatarUrl: "",
+  nickname: ["닉네임", ""],
+  level: ["레벨", 0],
+  winRate: ["승률", 0],
+  exp: ["경험치", 0],
+  isBanned: "false",
+  bannedReson: ""
+};
+
+const Profile = ({ isAdmin, profileInfos = init }) => {
   const [isPlaying, setIsPlaying] = useState(false);
   const [isIncludingKey, setIsIncludingKey] = useState(false);
   const [isActiveAccount, setIsActiveAccount] = useState(
@@ -49,7 +47,7 @@ const Profile = ({ isAdmin, profileInfos }) => {
             <AvatarImage src={profileInfos.avatarUrl} />
             <ProfileInfoWrapper dir="col" row="center">
               {isIncludingKey
-                ? Object.entries(profileInfos).map(([key, [title, value]], idx) => (
+                ? Object.entries(profileInfos)?.map(([key, [title, value]], idx) => (
                     <ProfileInfoField row="between" key={idx}>
                       <ProfileInfoKey>{title}</ProfileInfoKey>
                       <ProfileInfo>{value}</ProfileInfo>
@@ -57,7 +55,7 @@ const Profile = ({ isAdmin, profileInfos }) => {
                   ))
                 : Object.entries(profileInfos)
                     .filter(([key]) => ["nickname", "level"].includes(key))
-                    .map(([key, [title, value]], idx) => (
+                    ?.map(([key, [title, value]], idx) => (
                       <ProfileInfo key={idx}>
                         {key === "nickname" ? value : `${title} ${value}`}
                       </ProfileInfo>
@@ -93,17 +91,79 @@ Profile.propTypes = {
   profileInfos: PropTypes.object
 };
 
-Profile.defaultProps = {
-  isAdmin: false,
-  profileInfos: {
-    avatarUrl: "",
-    nickname: ["닉네임", ""],
-    level: ["레벨", 0],
-    winRate: ["승률", 0],
-    exp: ["경험치", 0],
-    isBanned: "false",
-    bannedReson: ""
+const ProfileWrapper = styled(FlexBox)``;
+
+const ProfileInfoWrapper = styled(FlexBox)`
+  margin-left: 10px;
+`;
+
+const ProfileInfoField = styled(FlexBox)`
+  width: 31.375rem;
+`;
+
+const ProfileInfoKey = styled.span``;
+
+const ProfileInfo = styled.span``;
+
+const AvatarImage = styled(FlexBox).attrs({
+  as: "img"
+})`
+  width: ${(props) => (props.type === "player" ? "7.75rem" : "6rem")};
+  height: ${(props) => (props.type === "player" ? "auto" : "5.304rem")};
+`;
+
+// ============ Exp ============
+const ExpWrapper = styled(FlexBox)`
+  width: ${(props) => props.width || "232px"};
+  height: 2.625rem;
+  margin-top: 16px;
+`;
+
+const ExpBar = styled.div`
+  width: 100%;
+  height: 100%;
+  border: 1px solid ${({ theme }) => theme.colors.gray600};
+  border-radius: 15px;
+`;
+
+const ProgressToNextLevel = styled.div`
+  width: ${({ exp }) => exp};
+  height: 100%;
+  background-color: ${({ theme }) => theme.colors.gray600};
+  border-radius: 13px 0 0 13px;
+`;
+
+const ExpText = styled.span`
+  position: relative;
+  left: 43%;
+  bottom: 33px;
+  color: #fff;
+  font-weight: 700;
+`;
+
+// ============ Active Toggle ============
+const ActiveToggle = styled(FlexBox)`
+  width: 75px;
+  height: 44px;
+  background-color: ${({ active }) => (active ? "#fff" : "#D9D9D9")};
+  border: 3px solid ${({ active }) => (active ? "#34E250" : "#787878")};
+  border-radius: 22px;
+
+  &:hover {
+    cursor: pointer;
   }
-};
+
+  & > div {
+    background-color: ${({ active }) => (active ? "#34E250" : "#787878")};
+    margin-left: ${({ active }) => !active && "3px"};
+    margin-right: ${({ active }) => active && "3px"};
+  }
+`;
+
+const ActiveCircle = styled.div`
+  width: 28px;
+  height: 28px;
+  border-radius: 50%;
+`;
 
 export default Profile;
