+++ src/pages/Web/Notice/NoticeList.jsx
@@ -1,11 +1,11 @@
 import React, { useState, useEffect } from "react";
 
 import { ContentWrapper, WideContent } from "@/styles/CommonStyle";
-import Header from "@/components/Shared/Layout/Header/Header";
-import Pagination from "@/components/Shared/Board/Pagination/Pagination";
+import Header from "@/components/Web/Shared/Layout/Header";
+import Pagination from "@/components/Web/Shared/Board/Pagination";
 
 const NoticeList = () => {
-  const [currPage, setCurrPage] = useState(1);
+  const [currPage, setCurrPage] = useState(16);
 
   useEffect(() => {}, [currPage]);
 
@@ -13,7 +13,11 @@ const NoticeList = () => {
     <ContentWrapper row="center" col="center">
       <WideContent dir="col">
         <Header type="detail" />
-        <Pagination currPage={currPage} setCurrPage={setCurrPage}></Pagination>
+        <Pagination
+          currPage={currPage}
+          setCurrPage={setCurrPage}
+          totalPageIdx="19"
+        ></Pagination>
       </WideContent>
     </ContentWrapper>
   );
